refactor(ui): extract the collapsible-header toggle into a hook
